requirements.txt
django
django-summernote
django-webpack-loader
djangorestframework
pillow
